refactor: 게임 조회 hook 각 파일로 분리
