+++ src/hooks/api/game/useGameListQuery.ts
@@ -1,0 +1,11 @@
+import { useQuery } from '@tanstack/react-query';
+import { GameProps, GamesPagination } from '@/types/game';
+import { getGameList } from '@/api/game';
+
+export const useGameList = (gameProps: GameProps) => {
+  const { data: gameList, isLoading } = useQuery<GamesPagination>({
+    queryKey: ['gameList', gameProps],
+    queryFn: () => getGameList(gameProps),
+  });
+  return { gameList, isLoading };
+};
